Remove unnecessary setContent calls in tests
